Refactor loot-core db queries into separate entity queries and mutations
