+++ packages/loot-core/src/server/db/queries/index.ts
@@ -1,0 +1,4 @@
+export * from './accounts';
+export * from './categories';
+export * from './payees';
+export * from './transactions';
